+++ Decoders/OggSpeexDecoder.h
@@ -1,7 +1,5 @@
 /*
- *  $Id$
- *
- *  Copyright (C) 2005 - 2007 Stephen F. Booth <me@sbooth.org>
+ *  Copyright (C) 2005 - 2020 Stephen F. Booth <me@sbooth.org>
  *
  *  This program is free software; you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -37,15 +35,15 @@
 	SpeexBits				_bits;
 	SpeexStereoState		_stereo;
 	
-	unsigned				_packetCount;
+	NSUInteger				_packetCount;
 	
-	unsigned				_framesPerPacket;
-	unsigned				_extraHeaderCount;
+	NSUInteger				_framesPerPacket;
+	NSUInteger				_extraHeaderCount;
 }
 
-- (unsigned)		packetCount;
+- (NSUInteger)		packetCount;
 
-- (unsigned)		framesPerPacket;
-- (unsigned)		extraHeaderCount;
+- (NSUInteger)		framesPerPacket;
+- (NSUInteger)		extraHeaderCount;
 
 @end
